feat: 커리큘럼 삭제 exception 추가

+++ src/curriculum/curriculum.resolver.ts
@@ -122,7 +122,14 @@ export class CurriculumResolver {
     @Args('input') input: DeleteCurriculumInput,
   ): Promise<boolean> {
     const { id } = input;
+
+    const curriculum = await this.curriculumService.getCurriculum({ id });
+    if (curriculum === null) {
+      throw new NotFoundException('curriculum not found');
+    }
+
     await this.curriculumService.deleteCurriculum(id);
+
     return true;
   }
